(refactor)(feat)Make webpack-dev-server is work properly with sockets
This auto reloads page so it will still disconnect from room on change.

We could change the disconnecting event for development, but we can
continue to update CSS directly in chrome with mapped files instead.

Hopefully it will make some aspects of development easier, but since
so much of the state is held by the server, it's not the case currently.